Fixed Play Registration Issue.
I added a req.session.email and req.session.name information to functions.js. So whenever someone registers their name and email will be the session ID. The session ID can change depending on the log in infromation. It was needed so that benchmarks could be added using the new Stats model.